basic app working
