+++ app/src/src-SessionBear/layouts/AppLayout/index.tsx
@@ -1,10 +1,7 @@
 import React, { useEffect } from "react";
-import { useLocation } from "react-router-dom";
-import APP_CONSTANTS from "../config/constants";
 import { submitAppDetailAttributes } from "utils/AnalyticsUtils.js";
 import { ConfigProvider } from "antd";
 import enUS from "antd/lib/locale/en_US";
-import FullScreenLayout from "layouts/FullScreenLayout";
 import ThirdPartyIntegrationsHandler from "hooks/ThirdPartyIntegrationsHandler";
 import { GrowthBookProvider } from "@growthbook/growthbook-react";
 import { growthbook } from "utils/feature-flag/growthbook";
@@ -18,14 +15,11 @@ import AutomationNotAllowedNotice from "components/misc/notices/AutomationNotAll
 import { useIsExtensionEnabled } from "hooks";
 import { LazyMotion, domMax } from "framer-motion";
 import ThemeProvider from "lib/design-system-v2/helpers/ThemeProvider";
-import DashboardLayout from "./layouts/DashboardLayout";
 import { useInitializeNewUserSessionRecordingConfig } from "features/settings/components/SessionsBookSettings/hooks";
 import DBListeners from "hooks/DbListenerInit/DBListeners";
+import { Outlet } from "react-router-dom";
 
-const { PATHS } = APP_CONSTANTS;
 const App = () => {
-  const location = useLocation();
-
   useEffect(() => {
     // Load features asynchronously when the app renders
     growthbook.loadFeatures({ autoRefresh: true });
@@ -43,7 +37,9 @@ const App = () => {
       <PreLoadRemover />
       <AppModeInitializer />
       <DBListeners />
+      {/* @ts-ignore */}
       <ActiveWorkspace />
+      {/* @ts-ignore */}
       <ThirdPartyIntegrationsHandler />
       <ThemeProvider>
         <ConfigProvider locale={enUS}>
@@ -51,13 +47,7 @@ const App = () => {
             <LocalUserAttributesHelperComponent />
             <LazyMotion features={domMax} strict>
               <div id="requestly-dashboard-layout">
-                {"/" + location.pathname.split("/")[1] === PATHS.LANDING ? (
-                  <FullScreenLayout />
-                ) : (
-                  <>
-                    <DashboardLayout />
-                  </>
-                )}
+                <Outlet />
               </div>
             </LazyMotion>
           </GrowthBookProvider>
